change investment project admin component to use existing project stage resource and remove newly created resource, refactor file structure moving state and tasks into existing files at the route of the projects folder inside investments
