--- src/client/components/Resource/InvestmentProjectStages.js
@@ -1,6 +1,0 @@
-import { createMetadataResource } from './Resource'
-
-export default createMetadataResource(
-  'InvestmentProjectStages',
-  'investment-project-stage'
-)
