+++ src/client/components/Resource/index.jsx
@@ -37,7 +37,6 @@ export { default as InvestmentProjectStageResource } from './ProjectStage'
 export { default as InvestmentTypesResource } from './InvestmentTypes'
 export { default as InvestorTypesResource } from './InvestorTypes'
 export { default as InvestmentInvestorTypesResource } from './InvestmentInvestorTypes'
-export { default as InvestmentProjectStagesResource } from './InvestmentProjectStages'
 export { default as InvestmentProjectTasksResource } from './InvestmentProjectTasks'
 export { default as LargeCapitalInvestmentEquityPercentagesResource } from './LargeCapitalInvestmentEquityPercentages'
 export { default as LargeCapitalInvestmentRestrictionsResource } from './LargeCapitalInvestmentRestrictions'
