--- src/client/modules/Investments/Projects/Admin/state.js
@@ -1,1 +1,0 @@
-export const TASK_UPDATE_STAGE = 'TASK_UPDATE_STAGE'
