+++ src/client/modules/Investments/Projects/state.js
@@ -15,6 +15,7 @@ export const TASK_EDIT_INVESTMENT_PROJECT_STATUS =
   'TASK_EDIT_INVESTMENT_PROJECT_STATUS'
 export const TASK_UPDATE_INVESTMENT_PROJECT_STAGE =
   'TASK_UPDATE_INVESTMENT_PROJECT_STAGE'
+export const TASK_UPDATE_STAGE = 'TASK_UPDATE_STAGE'
 
 export const INVESTMENT_PROJECTS_ID = 'projectsList'
 export const INVESTMENT_PROJECT_ID = 'investmentProject'
