+++ src/client/routes.js
@@ -94,7 +94,7 @@ import CompanyProjectsCollection from './modules/Companies/CompanyInvestments/Co
 import LargeCapitalProfile from './modules/Companies/CompanyInvestments/LargeCapitalProfile'
 import ExportsIndex from './modules/Companies/CompanyExports/ExportsIndex'
 import ExportsHistory from './modules/Companies/CompanyExports/ExportHistory'
-import InvestmentProjectAdmin from './modules/Investments/Projects/Admin/InvestmentProjectAdmin'
+import InvestmentProjectAdmin from './modules/Investments/Projects/InvestmentProjectAdmin'
 
 const routes = {
   companies: [
